Enhance Grok API integration and add environment template
