+++ .env.template
@@ -1,0 +1,22 @@
+# Environment variables for daily-podcast-stk
+# Copy this file to .env and fill in your secrets
+
+# xAI Grok
+GROK_API_URL=https://api.x.ai/v1/chat/completions
+GROK_API_KEY=
+# Legacy name (optional)
+XAI_API_KEY=
+
+# Kimi (Moonshot)
+MOONSHOT_API_KEY=
+
+# OpenRouter (optional)
+OPENROUTER_API_KEY=
+
+# Backblaze B2
+B2_KEY_ID=
+B2_APPLICATION_KEY=
+B2_BUCKET_NAME=daily-podcast-stk
+
+# Podcast mode override (us|tw). Usually set by workflow
+PODCAST_MODE=
