+++ grok_api.py
@@ -3,6 +3,12 @@ import os
 from openai import OpenAI
 from time import sleep
 from requests.exceptions import RequestException
+import requests
+import json
+import re
+
+GROK_API_URL = os.getenv("GROK_API_URL", "https://api.x.ai/v1/chat/completions")
+GROK_API_KEY = os.getenv("GROK_API_KEY") or os.getenv("XAI_API_KEY")
 
 def optimize_script_with_grok(initial_script, api_key, model="grok-4", max_retries=3):
     if not api_key:
@@ -45,3 +51,41 @@ def optimize_script_with_grok(initial_script, api_key, model="grok-4", max_retri
             continue
     print(f"Grok API 調用失敗 {max_retries} 次，使用初始逐字稿")
     return initial_script
+
+def ask_grok(prompt: str, role: str = "user", model: str = "grok-4") -> str:
+    """Call xAI Grok API and return plain text reply."""
+    if not GROK_API_KEY:
+        raise EnvironmentError("❌ 請設定 GROK_API_KEY 或 XAI_API_KEY 環境變數")
+
+    headers = {
+        "Authorization": f"Bearer {GROK_API_KEY}",
+        "Content-Type": "application/json",
+    }
+    payload = {
+        "model": model,
+        "messages": [{"role": role, "content": prompt}],
+        "temperature": 0.7,
+        "max_tokens": 2048,
+        "stream": False,
+    }
+
+    resp = requests.post(GROK_API_URL, headers=headers, json=payload, timeout=60)
+    resp.raise_for_status()
+    data = resp.json()
+    choices = data.get("choices", [])
+    if not choices or not choices[0].get("message", {}).get("content"):
+        raise RuntimeError(f"❌ Grok 回傳內容為空或無效：{json.dumps(data, ensure_ascii=False)[:500]}")
+    return choices[0]["message"]["content"].strip()
+
+
+def ask_grok_json(prompt: str, role: str = "user", model: str = "grok-4") -> dict:
+    """Call xAI Grok API and parse a JSON object from the reply."""
+    reply = ask_grok(prompt, role=role, model=model)
+    json_match = re.search(r"\{[\s\S]*\}", reply, re.DOTALL)
+    if not json_match:
+        raise ValueError(f"❌ Grok 回傳不是合法 JSON：\n{reply[:1000]}")
+    json_str = json_match.group(0)
+    json_str = re.sub(r"(\d+)\.(?!\d)", r"\1.0", json_str)
+    json_str = re.sub(r",\s*}", r"}", json_str)
+    json_str = re.sub(r",\s*,", r",", json_str)
+    return json.loads(json_str)
