+++ requirements.txt
@@ -17,5 +17,4 @@ boto3
 #vectorbt
 twstock
 numba==0.58.1
-lxml
 openai
